refactor: Move the kernel post init stage into proc0 refactor: Add a linter config. Resolved #1 refactor: Formatted some source files. refactor: Shrink the process table size.

+++ lunaix-os/includes/lunaix/lunistd.h
@@ -8,7 +8,7 @@ __LXSYSCALL(pid_t, fork)
 
 __LXSYSCALL1(int, sbrk, void*, addr)
 
-__LXSYSCALL1(void*, brk, size_t, size)
+__LXSYSCALL1(void*, brk, unsigned long, size)
 
 __LXSYSCALL(pid_t, getpid)
